+++ workflow/rules/ref.smk
@@ -72,28 +72,42 @@ rule bed2gff:
         "../envs/genometools.yaml"
     shell:
         "(cut -f1-12 {input} | sed -e 's/ /-/g' | sed -e 's/NC_045512v2/NC_045512.2/g'"
-        " | gt bed_to_gff3 -featuretype gene -thicktype transcript -blocktype exon -o {output} -force /dev/stdin )"
+        " | gt bed_to_gff3 -featuretype gene -thicktype transcript -blocktype CDS -o {output} -force /dev/stdin )"
         "2> {log}"
 
 
-# rule sort_gff:
-#     input:
-#         "resources/protein_products.gff",
-#     output:
-#         temp("resources/protein_products.sorted.gff"),
-#     log:
-#         "logs/sort_gff.log"
-#     conda:
-#         "../envs/gff3sort.yaml"
-#     shell:
-#         "gff3sort.pl --precise --chr_order natural {input} > {output}"
-
-
-rule format_gff:
+rule filter_gff:
     input:
         "resources/protein_products.gff",
     output:
+        temp("resources/protein_products.formatted.gff"),
+    log:
+        "logs/format_gff.log",
+    conda:
+        "../envs/tabix.yaml"
+    shell:
+        # download, sort and bgzip gff (see https://www.ensembl.org/info/docs/tools/vep/script/vep_custom.html)
+        "cat {input} | grep -v '#' > {output} 2> {log}"
+
+
+rule fix_gff:
+    input:
         "resources/protein_products.formatted.gff",
+    output:
+        temp("resources/protein_products.fixed.gff"),
+    log:
+        "logs/fix_gff.log",
+    conda:
+        "../envs/python.yaml"
+    script:
+        "../scripts/fix-protein-gff.py"
+
+
+rule format_fixed_gff:
+    input:
+        "resources/protein_products.fixed.gff",
+    output:
+        temp("resources/protein_products.fixed.formatted.gff"),
     log:
         "logs/format_gff.log",
     conda:
@@ -103,22 +117,9 @@ rule format_gff:
         "cat {input} | grep -v '#' | sort -k1,1 -k4,4n -k5,5n -k3,3n -t$'\t' > {output} 2> {log}"
 
 
-rule fix_gff:
-    input:
-        "resources/protein_products.formatted.gff",
-    output:
-        "resources/protein_products.fixed.gff",
-    log:
-        "logs/fix_gff.log",
-    conda:
-        "../envs/python.yaml"
-    script:
-        "../scripts/fix-gff.py"
-
-
 rule compress_gff:
     input:
-        "resources/protein_products.fixed.gff",
+        "resources/protein_products.fixed.formatted.gff",
     output:
         "resources/protein_products.gff.gz",
     log:
@@ -127,33 +128,6 @@ rule compress_gff:
         "../envs/tabix.yaml"
     shell:
         "bgzip -c {input} > {output} 2> {log}"
-
-
-# rule get_protein_products:
-#     input:
-#         "resources/protein_products.bed.gz",
-#     output:
-#         "resources/protein_products.gft.gz",
-#     log:
-#         "logs/get_protein_products.log"
-#     conda:
-#         "../envs/ucsc.yaml"
-#     shell:
-#         """
-#         (set +e
-#             ( bgzip -dkc {input} | \
-#             bedToGenePred /dev/stdin /dev/stdout | \
-#             genePredToGtf file -source=ucsc /dev/stdin /dev/stdout | \
-#             bgzip -c > {output})
-#             exitcode=$?
-#             echo 'exitcode:' $exitcode >&2
-#             if [ $exitcode -eq 1 ]
-#             then
-#                 exit 1
-#             else
-#                 exit 0
-#             fi) 2> {log}
-#         """
 
 
 rule get_genome_annotation_for_known_variants:
